Fixed issue with highlighting the beginning of functions
See: https://github.com/snare/binjatron/issues/1

`bv.get_function_at(address)` only works if it's called on the very first address of a function. `bv.get_previous_function_start(address)` typically returns the address of the function containing an address, but if that address is the very first one in the function, it'll go back to the previous function in the linear disassembly. This fix was as simple as creating a wrapper that checks whether we're at the beginning of a function before returning.