Fixed lingering instruction pointer highlights on call instructions
When a call (or jump outside of a function) is made, binary ninja no longer returns the same function for the current program counter as it would for `last_pc_addr`. Since the two functions don't match, when we try to remove the highlighting using the wrong one, it fails silently.